Moved utils to utiles folder
